refactor: common translation functions
