+++ src/domain/entities/TranslatableText.ts
@@ -16,3 +16,18 @@ export const buildTranslate = (locale: string): TranslateMethod => {
 };
 
 export type TranslateMethod = (string: TranslatableText) => string;
+
+type TranslateProps<T> = {
+    item: T;
+    language: string;
+    term: string | undefined;
+};
+export function setTranslationValue<T extends TranslatableText>({ item, language, term }: TranslateProps<T>): T {
+    if (term === undefined) {
+        return item;
+    } else if (language === "en") {
+        return { ...item, referenceValue: term };
+    } else {
+        return { ...item, translations: { ...item.translations, [language]: term } };
+    }
+}
