+++ src/domain/usecases/ImportLandingNodesTranslationsUseCase.ts
@@ -1,34 +1,28 @@
 import _ from "lodash";
 import { PersistedLandingNode } from "../../data/entities/PersistedLandingNode";
-import { TranslatableText } from "../entities/TranslatableText";
+import { setTranslationValue } from "../entities/TranslatableText";
 import { LandingNodeRepository } from "../repositories/LandingNodeRepository";
 import { UseCase } from "./UseCase";
-import { extractTranslations } from "./common";
+import { TranslationService } from "./helpers/TranslationService";
 
 export class ImportLandingNodesTranslationsUseCase implements UseCase {
     constructor(private landingNodeRepository: LandingNodeRepository) {}
 
-    public async execute(lang: string, terms: Record<string, string>, key: string): Promise<number> {
+    public async execute(language: string, terms: Record<string, string>, key: string): Promise<number> {
         const persisted = await this.landingNodeRepository.getPersistedLandingPages();
-
-        const translate = <T extends TranslatableText>(item: T, language: string, term: string | undefined): T => {
-            if (term === undefined) {
-                return item;
-            } else if (language === "en") {
-                return { ...item, referenceValue: term };
-            } else {
-                return { ...item, translations: { ...item.translations, [language]: term } };
-            }
-        };
 
         const toTranslate = persisted.find(model => model.find(item => item.id === key));
         if (!toTranslate) throw new Error(`Unable to load landing pages`);
 
         const translatedModels: PersistedLandingNode[] = toTranslate.map(model => ({
             ...model,
-            name: translate(model.name, lang, terms[model.name.key]),
-            title: model.title ? translate(model.title, lang, terms[model.title.key]) : undefined,
-            content: model.content ? translate(model.content, lang, terms[model.content.key]) : undefined,
+            name: setTranslationValue({ item: model.name, language, term: terms[model.name.key] }),
+            title: model.title
+                ? setTranslationValue({ item: model.title, language, term: terms[model.title.key] })
+                : undefined,
+            content: model.content
+                ? setTranslationValue({ item: model.content, language, term: terms[model.content.key] })
+                : undefined,
         }));
 
         const updatedLandingNodes = persisted.map(model => {
@@ -38,7 +32,7 @@ export class ImportLandingNodesTranslationsUseCase implements UseCase {
 
         await this.landingNodeRepository.save(updatedLandingNodes);
 
-        const translations = await extractTranslations(translatedModels);
+        const translations = TranslationService.extractTranslations(translatedModels);
 
         return _.intersection(_.keys(translations["en"]), _.keys(terms)).length;
     }
